requirements.txt
langchain
langchain-groq
langchain-community
faiss-cpu
pypdf
streamlit
python-dotenv
sentence-transformers